refactor: all fitting functions are done!
